+++ app/src/main/java/com/example/filmsapp/ui/details/DetailsViewModel.kt
@@ -12,10 +12,7 @@ import com.example.filmsapp.ui.base.Event
 import com.example.filmsapp.ui.base.models.FilmModel
 import com.example.filmsapp.ui.base.models.YoutubeFilmModel
 import com.google.api.client.googleapis.extensions.android.gms.auth.GoogleAccountCredential
-import com.google.api.client.googleapis.extensions.android.gms.auth.GooglePlayServicesAvailabilityIOException
-import com.google.api.client.googleapis.extensions.android.gms.auth.UserRecoverableAuthIOException
 import kotlinx.coroutines.launch
-import timber.log.Timber
 
 class DetailsViewModel(
     dispatcherProvider: DispatcherProvider,
@@ -28,12 +25,6 @@ class DetailsViewModel(
 
     private val _isFavorites = MutableLiveData<Event<Boolean>>()
     val isFavorites: LiveData<Event<Boolean>> get() = _isFavorites
-
-    private val _requestAuthorizationPermission = MutableLiveData<UserRecoverableAuthIOException>()
-    val requestAuthorizationPermission: LiveData<UserRecoverableAuthIOException> get() = _requestAuthorizationPermission
-
-    private val _displayGpsUnavailable = MutableLiveData<Int>()
-    val displayGpsUnavailable: LiveData<Int> get() = _displayGpsUnavailable
 
     private val _youtubeMovieSearchResult = MutableLiveData<YoutubeFilmModel>()
     val youtubeMovieSearchResult: LiveData<YoutubeFilmModel> get() = _youtubeMovieSearchResult
@@ -88,19 +79,5 @@ class DetailsViewModel(
                 postMessage(R.string.film_removed)
             }
         } ?: postMessage(R.string.error)
-    }
-
-    override fun handleException(exception: Throwable) {
-        super.handleException(exception)
-        when (exception) {
-            is GooglePlayServicesAvailabilityIOException ->
-                _displayGpsUnavailable.value =
-                    exception.connectionStatusCode
-            is UserRecoverableAuthIOException -> _requestAuthorizationPermission.value = exception
-            else -> {
-                Timber.e(exception)
-                postMessage(R.string.error)
-            }
-        }
     }
 }
