+++ app/src/main/res/navigation/root.xml
@@ -2,7 +2,7 @@
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/root"
-    app:startDestination="@id/listsFragment">
+    app:startDestination="@id/splashFragment">
 
     <fragment
         android:id="@+id/mainFragment"
@@ -79,5 +79,13 @@
         android:id="@+id/settingsFragment"
         android:name="com.example.filmsapp.ui.settings.SettingsFragment"
         android:label="SettingsFragment" />
+    <fragment
+        android:id="@+id/splashFragment"
+        android:name="com.example.filmsapp.ui.splash.SplashFragment"
+        android:label="SplashFragment" >
+        <action
+            android:id="@+id/action_splashFragment_to_listsFragment"
+            app:destination="@id/listsFragment" />
+    </fragment>
 
 </navigation>
